@@ -1,12 +1,12 @@
 apply plugin: 'com.android.library'
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 29
     buildToolsVersion "28.0.3"
 
     defaultConfig {
         minSdkVersion 21
-        targetSdkVersion 28
+        targetSdkVersion 29
         versionCode 1
         versionName "1.0"
 
@@ -16,10 +16,9 @@ android {
 
 dependencies {
     api fileTree(dir: "libs", include: ["*.jar"])
-//    api 'com.android.support:appcompat-v7:28.0.0'
-    api 'com.android.support:palette-v7:28.0.0'
-    api 'com.android.support.constraint:constraint-layout:1.1.3'
+    api 'androidx.palette:palette:1.0.0'
+    api 'androidx.constraintlayout:constraintlayout:2.0.4'
 
     //https://github.com/actor20170211030627/MyAndroidFrameWork
-    api 'com.github.actor20170211030627.MyAndroidFrameWork:myandroidframework:1.4.1'
+    api 'com.github.actor20170211030627.MyAndroidFrameWork:myandroidframework:2.0.1'
 }
