@@ -7,14 +7,15 @@ import android.graphics.Path;
 import android.graphics.drawable.Animatable;
 import android.graphics.drawable.Drawable;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewAnimationUtils;
 import android.view.ViewGroup;
 import android.view.animation.LinearInterpolator;
 import android.view.animation.PathInterpolator;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import com.develophelper.android5.R;
 import com.develophelper.android5.activity.TransitionsActivity;
@@ -96,6 +97,7 @@ public class AnimationFragment extends BaseFragment implements View.OnClickListe
         mAnimator.start();
     }
 
+    @Override
     public String getUrl() {
         return "file:///android_asset/animation.html";
     }
