@@ -4,13 +4,14 @@ import android.graphics.Color;
 import android.graphics.drawable.BitmapDrawable;
 import android.graphics.drawable.Drawable;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v7.graphics.Palette;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.palette.graphics.Palette;
 
 import com.develophelper.android5.R;
 
@@ -57,6 +58,7 @@ public class DrawableFragment extends BaseFragment {
         return view;
     }
 
+    @Override
     public String getUrl() {
         return "file:///android_asset/drawable.html";
     }
