@@ -6,9 +6,10 @@ import android.app.ProgressDialog;
 import android.app.Service;
 import android.content.Intent;
 import android.os.IBinder;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
 import android.view.WindowManager;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import com.actor.androiddevelophelper.Global;
 import com.actor.androiddevelophelper.info.CheckUpdateInfo;
